+++ src/hdmf/backends/hdf5/h5tools.py
@@ -311,10 +311,10 @@ class HDF5IO(HDMFIO):
              'doc': 'A HERD object to populate with references.',
              'default': None},
             {'name': 'expandable', 'type': bool, 'default': None,
-             'doc': ('If None (default), only VectorData, ElementIdentifiers, and their subclasses datasets '
-                     'will be created as expandable. If True, all datasets will be created as expandable. '
-                     'If False, no datasets will be automatically made expandable. Expandable datasets have '
-                     'maxshape set based on the matching shape defined in the spec.')})
+             'doc': ('If None (default), only datasets of type VectorData, ElementIdentifiers, or their '
+                     'subclasses will be created as expandable. If True, all datasets will be created as '
+                     'expandable. If False, no datasets will be automatically made expandable. Expandable '
+                     'datasets have maxshape set based on the matching shape defined in the spec.')})
     def write(self, **kwargs):
         """Write the container to an HDF5 file."""
         if self.__mode == 'r':
@@ -762,10 +762,10 @@ class HDF5IO(HDMFIO):
             {'name': 'export_source', 'type': str,
              'doc': 'The source of the builders when exporting', 'default': None},
             {'name': 'expandable', 'type': bool, 'default': None,
-             'doc': ('If None (default), only VectorData, ElementIdentifiers, and their subclasses datasets '
-                     'will be created as expandable. If True, all datasets will be created as expandable. '
-                     'If False, no datasets will be automatically made expandable. Expandable datasets have '
-                     'maxshape set based on the matching shape defined in the spec.')})
+             'doc': ('If None (default), only datasets of type VectorData, ElementIdentifiers, or their '
+                     'subclasses will be created as expandable. If True, all datasets will be created as '
+                     'expandable. If False, no datasets will be automatically made expandable. Expandable '
+                     'datasets have maxshape set based on the matching shape defined in the spec.')})
     def write_builder(self, **kwargs):
         f_builder = popargs('builder', kwargs)
         self.logger.debug("Writing GroupBuilder '%s' to path '%s' with kwargs=%s"
@@ -944,10 +944,10 @@ class HDF5IO(HDMFIO):
             {'name': 'export_source', 'type': str,
              'doc': 'The source of the builders when exporting', 'default': None},
             {'name': 'expandable', 'type': bool, 'default': None,
-             'doc': ('If None (default), only VectorData, ElementIdentifiers, and their subclasses datasets '
-                     'will be created as expandable. If True, all datasets will be created as expandable. '
-                     'If False, no datasets will be automatically made expandable. Expandable datasets have '
-                     'maxshape set based on the matching shape defined in the spec.')},
+             'doc': ('If None (default), only datasets of type VectorData, ElementIdentifiers, or their '
+                     'subclasses will be created as expandable. If True, all datasets will be created as '
+                     'expandable. If False, no datasets will be automatically made expandable. Expandable '
+                     'datasets have maxshape set based on the matching shape defined in the spec.')},
             returns='the Group that was created', rtype=Group)
     def write_group(self, **kwargs):
         parent, builder = popargs('parent', 'builder', kwargs)
@@ -1049,10 +1049,10 @@ class HDF5IO(HDMFIO):
             {'name': 'export_source', 'type': str,
              'doc': 'The source of the builders when exporting', 'default': None},
             {'name': 'expandable', 'type': bool, 'default': None,
-             'doc': ('If None (default), only VectorData, ElementIdentifiers, and their subclasses datasets '
-                     'will be created as expandable. If True, all datasets will be created as expandable. '
-                     'If False, no datasets will be automatically made expandable. Expandable datasets have '
-                     'maxshape set based on the matching shape defined in the spec.')},
+             'doc': ('If None (default), only datasets of type VectorData, ElementIdentifiers, or their '
+                     'subclasses will be created as expandable. If True, all datasets will be created as '
+                     'expandable. If False, no datasets will be automatically made expandable. Expandable '
+                     'datasets have maxshape set based on the matching shape defined in the spec.')},
             returns='the Dataset that was created', rtype=Dataset)
     def write_dataset(self, **kwargs):  # noqa: C901
         """ Write a dataset to HDF5